move frep instructions to riscv dialect file
